track2_finetuning/requirements.txt
# PyTorch + ROCm is installed via setup_rocm.sh (pip index URL), not here.
# Leave torch/torchvision OUT of this file so a stray `pip install -r` on a CUDA
# box doesn't overwrite the ROCm wheels.

timm>=1.0.9
numpy>=1.26
pillow>=10.0
pyyaml>=6.0
scikit-learn>=1.3
tqdm>=4.66
tensorboard>=2.15
pandas>=2.1
